moved the file and added readme
